+++ .github/workflows/build.yml
@@ -46,10 +46,11 @@ jobs:
         run: |
           python -m pip install -U pip
           python -m pip install -U setuptools wheel
-          python -m pip install -U coverage coveralls
+          python -m pip install -U coverage coveralls pytest
+          python -m pip install -e .
 
       - name: Run tests
-        run: coverage run testsuite.py
+        run: coverage run -m pytest
 
       - name: Check test coverage
         run: |
@@ -69,6 +70,7 @@ jobs:
       matrix:
         toxenv:
           - flake8
+          - ruff
           - isort
           - check-manifest
           - check-python-versions
