+++ MANIFEST.in
@@ -1,17 +1,14 @@
+include *.mk
 include *.rst
+include *.yml
+include LICENSE
 include Makefile
 include .coveragerc
 include .gitignore
 include tox.ini
-include testsuite.py
+include pytest.ini
+include conftest.py
 include tests.py
 include tests/*.txt
 include tests/sample-tree/snake.egg-info
 recursive-include tests/sample-tree *.py *.zip
-
-# added by check_manifest.py
-include *.yml
-include LICENSE
-
-# added by check_manifest.py
-include *.mk
