+++ pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+testpaths = tests.py tests/*.txt
+addopts = -ra -s
