+++ tests/cyclic-imports.txt
@@ -1,7 +1,5 @@
 Test cyclic imports
 
-    >>> from testsuite import create_tree
-
     >>> create_tree('''
     ... -- foo/__init__.py --
     ...
