+++ tox.ini
@@ -7,30 +7,38 @@ envlist =
     py314
     pypy3
     flake8
+    ruff
     isort
     check-python-versions
 
 [testenv]
 deps =
+    pytest
 commands =
-    python testsuite.py {posargs}
+    pytest {posargs}
 
 [testenv:coverage]
 deps =
+    pytest
     coverage
 commands =
-    coverage run testsuite.py
+    coverage run -m pytest
     coverage report -m --fail-under=100
 
 [testenv:flake8]
 deps = flake8
 skip_install = true
-commands = flake8 findimports.py setup.py tests.py testsuite.py
+commands = flake8 findimports.py setup.py tests.py conftest.py
+
+[testenv:ruff]
+deps = ruff
+skip_install = true
+commands = ruff check findimports.py setup.py tests.py conftest.py
 
 [testenv:isort]
 deps = isort
 skip_install = true
-commands = isort {posargs: -c --diff findimports.py setup.py tests.py testsuite.py}
+commands = isort {posargs: -c --diff findimports.py setup.py tests.py conftest.py}
 
 [testenv:check-manifest]
 deps = check-manifest
